Filter aborted SNSes from snsAggregatorStore (#5552)
# Motivation

We had a bug because we tried to parse nervous system parameters on
aborted SNSes.
All live SNSes do have nervous system parameters and we don't display
aborted SNSes so there is no reason to parse them.

# Changes

1. Rename `snsAggregatorStore` to
`snsAggregatorIncludingAbortedProjectsStore` to avoid confusion and add
`snsAggregatorStore` as a derived store which does not contain aborted
SNSes.

# Tests

1. Update sets which (re)set data on `snsAggregatorStore` to set it on
`snsAggregatorIncludingAbortedProjectsStore`.
2. Add a utility to test with prod SNS data.
3. Updated `setSnsProjects` to allow setting `nervousSystemParameters:
null`.
4. Add tests that `SnsProposalDetail` can render with aborted SNSes
(explicitly) and with prod SNSes (which also contain aborted SNSes).
5. Add unit tests for new derived `snsAggregatorStore`.
6. Test that stores derived from `snsAggregatorStore` can be subscribed
with prod data.
7. Tested manually against mainnet.

# Todos

- [x] Add entry to changelog (if necessary).