Remove separate uploading of assets from deploy-to-app.yaml (#5558)
# Motivation

The nns-dapp wasm is bigger than 2 MB. When the `deploy-to-app` workflow
was created, this meant that the wasm couldn't be installed to an
application subnet.
The "real" nns-dapp isn't affected by this because the NNS subnet allows
larger messages.
To work around this, a mechanism was implemented to build and install a
wasm without assets and upload the assets separately.
This is no longer necessary because the IC now supports uploading the
wasm in chunks and then installing the wasm from chunks. `dfx` also
supports this transparently.
So we can simply call `dfx` to install the large wasm and everything
happens correctly behind the scenes.

# Changes

1. Remove the calls to split the assets.
2. Install the standard wasm instead of the `noassets` wasm.
3. Remove the calls to upload the assets.

# Tests

https://beta.nns.ic0.app/ is working again because it was installed and
upgraded this way:
* reinstall:
https://github.com/dfinity/nns-dapp/actions/runs/11125543897
* upgrade: https://github.com/dfinity/nns-dapp/actions/runs/11125814854

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary