Fix uncertified error in loadIcrcToken (#5565)
# Motivation

In `loadIcrcToken` we try to ignore errors on query calls if there is
also an update call.
For this we use the following logic:
```
      if (!certified && notForceCallStrategy()) {
        return;
      }
```
This assumes that we make an update call unless `notForceCallStrategy()`
returns `false`.

But it's also possible to pass `certified: false` into `loadIcrcToken`
and then we also don't make an update call, but the error is still
skipped on query calls so in that case we ignore all errors.

# Changes

1. Take the `certified` that's passed `loadIcrcToken` into into account
to determine whether to ignore errors.

# Tests

1. Add unit tests for `loadIcrcToken`.
2. Mock `queryIcrcToken` in `workflows/Launchpad/ProdLaunchpad.spec.ts`
as it is now showing errors that were always there about not being able
to destructure its result.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary