Fix replacement of neuron ID in screenshot test (#5572)
# Motivation

In the neuron details e2e test we replace the neuron ID with a fixed
value so that it doesn't cause the screenshot to be different each time.
When doing this we expect the existing neuron to have 19 digits, but
neuron IDs can also have fewer than 19 digits.
And when the existing neuron ID doesn't match the expected pattern it's
not replaced which results in test flakiness.

# Changes

1. Relax the expected neuron ID pattern.
2. Expect the number of replacements for each call to `replaceContent`
to be greater than 0 to help catch cases where nothing is replaced
without having to download the screenshot.

# Tests

Only tests.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary