Fix mistakenly introduced code changes (#5574)
# Motivation

NnsNeuronAge has label but all the other keyValue pairs don't. This was
introduced by mistake in [here
](https://github.com/dfinity/nns-dapp/pull/5516/files#diff-8fdc14d15bb871489f65e3413e08042acdf48c6eed95ecea2ce5cacd745b705bR23).
There was a time all the key value pairs had the correct color according
to Figma in referenced pr. When reverting style changes, I forgot to
delete this one. It looks like it passed by in between all other code
changes as this looks like a small style update.

As you can see in the screenshots it is also not easily noticeable.

# Changes

Delete wrong styling for NnsNeuronAge

# Tests

screenshot tests

# Todos

- [ ] Add entry to changelog (if necessary).