Download ledger.did from new path (#5617)
# Motivation

"Update sns_aggregator candid bindings"
[failed](https://github.com/dfinity/nns-dapp/actions/runs/11338847183/job/31540786783)
because ledger and Index Candid files were moved to a different
directory in https://github.com/dfinity/ic/pull/1682

After fixing the path in the script, I had to also update the candid
files to the new commit, because we have a
[test](https://github.com/dfinity/nns-dapp/blob/a360d0281792040f8930f089c59aef9996188024/.github/workflows/checks.yml#L140-L143)
that checks that the current code matches what would be generated at the
current version in `dfx.json` and without updating `dfx.json` it would
try to download from the new path at the old version which also fails.

The important changes are in the first commit and updating the candid
and bindings is in the second commit.

# Changes

1. Change the path where we download the `ledger.did` to the new path.
2. Add `-S` to `curl` to get slightly better error messages.
3. Ran `scripts/update_ic_commit --crate sns_aggregator --ic_commit
"release-2024-10-11_14-35-overload"`.
4. Ran `scripts/sns/aggregator/mk_nns_types.sh`.

# Tests

Ran `scripts/update_ic_commit --crate sns_aggregator --ic_commit
"release-2024-10-11_14-35-overload"` with and without the new changes
and saw it first fail and then pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary