Add checkbox to visibility row (#5631)
# Motivation

This is a breakdown of [large
pr](https://github.com/dfinity/nns-dapp/pull/5606), which initially
holds the checkbox for neruons in it the ChangeBulkNeuronVisibilityForm
component.

NeuronVisibilityRow according to design is mainly a checkbox and only
displaying label part of the row as NeuronVisibilityRow doesn't make
sense. We most probably want the row to hold all elements in single row,
that is why i would like move checkbox into NeuronVisibilityRow.

# Changes

Add a `Checkbox` to the `NeuronVisibilityRow` component.

# Tests

Tests are added to make sure that checkbox functions as expected.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary