Get canister ID with dfx in e2e test (#5647)
# Motivation

The ledger and index canister of ckRED are hard coded in
`frontend/src/tests/e2e/import-token.spec.ts`.
But the canister IDs are not guaranteed to remain the same in snsdemo.
In particular, with the latest snapshot, they are different, resulting
in test failures in the [automatic update
PR](https://github.com/dfinity/nns-dapp/pull/5645).

# Changes

1. Add a utility function that calls `dfx` to get a canister ID.
2. Use the new function to get the correct canister IDs in the e2e test.

# Tests

The test passes locally with a newer snapshot and still passes on
current CI as well.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary