Remove menu footer when menu is collapsed (#5649)
# Motivation

When the menu is collapsed, the menu footer (with TVL and code/theme
buttons) is moved out of view, but it still takes up space in the DOM.
This means the scrollbar appears sooner than necessary on small screens.

# Changes

1. Use `{#if}` to remove the menu footer when the menu is collapsed.
2. Use Svelte animation to animate the dis/reappearance instead of CSS
transition. Use `cubicIn` and `cubicOut` because they make sure the
element is always smaller than the space it has, which avoids strange
looking reflowing of the content.

# Tests

1. Unit tests added.
2. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [x] Add entry to changelog (if necessary).