Don't expect specific canister IDs in snsdemo snapshot (#5650)
# Motivation

We have a golden canister IDs file for the local network to test that
importing canister IDs works.
But canister IDs in the snsdemo snapshot can change.
This caused a [test
failure](https://github.com/dfinity/nns-dapp/actions/runs/11399975435/job/31720267156?pr=5645)
in the snsdemo [update
PR](https://github.com/dfinity/nns-dapp/pull/5645).

# Changes

1. Instead of comparing the `canister_ids.json` file for the local
network to a golden file, check that the expected keys are present. We
still have a golden file for the mainnet network so we do still test the
file structure of the file as well.
2. Delete `scripts/canister_ids.local.golden`.

# Tests

CI still passes.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary