Fix outdated rerender syntax (#5642)
# Motivation

Upgrading the test dependencies will break some tests. To mitigate this,
we are improving our tests to make them less dependent on the test
library versions.

# Changes

- Replace rerender in Settings.spec with runResolvedPromises because the
tested value is based on the authRemainingTimeStore state, and the
component does not need to be fully rerendered to display the correct
information.
- Update the SnsProposalDetail test to use $set instead of rerender, to
avoid syntax change after the testing library upgrade. The test has been
verified to fail when [the fix it
tests](https://github.com/dfinity/nns-dapp/pull/2663) is reverted.

# Tests

Yes.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary