Use real toastsStore in vote-registration.services.spec.ts (#5662)
# Motivation

Making the test more robust by using the real `toastsStore` instead of
mocking the calls the update the store.

Note: In the process of cleaning up the test I also found 2 bugs:
1. When showing the toasts which say "Neurons registered: x/y.", the
proposal type that's displayed when x == 0 is actually the proposal
topic instead of the type.
2. The test `"should display updating... message"` no longer appears to
be testing what it should be testing because its expectations inside the
for loop do not depend on the loop variable.

I did not fix these bugs but I did comment on the corresponding tests.
The bugs can be fixed in separate PRs.

# Changes

1. Stop mocking `spyOnToastsUpdate`, `spyOnToastsShow`, and
`spyOnToastsError` and instead inspect the contents of the actual
`toastsStore`.
2. Expect the store to be empty before the code under test is called.
3. Expect literal string messages instead of objects with keys.
4. Because the toasts are also replaced in the code, some tests provide
a mock implementation for `registerVote` or `loadProposal` which doesn't
resolve in order to be able to see a toast before it's replaced. To make
this work `loadProposal` is no longer mocked at the file level but with
`spyOn`.

# Tests

Test only change.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary