NNS1-3391: Add error messages for incorrect transaction timestamps (#5667)
# Motivation

There was a [forum
post](https://forum.dfinity.org/t/i-having-trouble-transferring-icp-from-the-nns-dapp/36038)
from a user who kept getting errors when trying to transfer ICP.
In the end it turned out to be a problem with their computer's time
settings.

To avoid this situation in the future we want to provide better error
messages.

# Changes

1. Remove the `labelKey` parameter from `transferError` as
`transferError` is only called in 1 place with a hardcoded `labelKey`
argument.`
2. In `transferError` use the originally passed-in `labelKey` as default
but use a different `labelKey` if the error is `TxTooOldError` or
`TxCreatedInFutureError`.
3. Add new messages for the new label keys.

# Tests

1. Add missing test coverage for a generic error from `sendICP`.
2. Add tests for `TxTooOldError` and `TxCreatedInFutureError`.
3. I was unable to reproduce the issue manually because whenever I
changed my clock settings, either nothing would change or I would got
very severe errors before I could even try to make a transaction. So I
just hardcoded the relevant errors in my local `ic-js` repo to test this
manually.

# Todos

- [x] Add entry to changelog (if necessary).