Show proposal type in vote registration (#5670)
# Motivation

While voting, we update the toast multiple times to show the status of
the voting.
The toast also shows the proposal ID and topic or type. But whether it
shows topic or type is inconsistent between toasts.

# Changes

Consistently identify the proposal by its topic in vote registration
toasts.

# Tests

Tests are updated.

# Todos

- [x] Add entry to changelog (if necessary).