display default selected neuron (#5669)
# Motivation

To give users more confidence on the action they are taking we would
like to display the default selected neuron, also it would be nice that
ChangeNeuronVisibilityModal and ChangeBulkNeuronVisibilityForm don't
depend on neuron that is passed to function so we can use it directly in
the banner.

# Changes

1. Update ChangeBulkNeuronVisibilityForm ChangeNeuronVisibilityModal to
have defaultSelectedNeuron and isPublic props instead of neuron
2. Update NnsNeuronModals to pass the new props

# Tests

ChangeBulkNeuronVisibilityForm and ChangeNeuronVisibilityModal tests
have been updated.

The default selected neuron comes selected and listed in the Neurons
list:
<img width="621" alt="Screenshot 2024-10-22 at 18 05 10"
src="https://github.com/user-attachments/assets/bb646000-847d-48f9-971d-15ca84655794">

# Todos

- [ ] Add entry to changelog (if necessary).

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: “Cosku <“cosku.cinkilic@dfinity.org”>
Co-authored-by: sa-github-api <138766536+sa-github-api@users.noreply.github.com>
Co-authored-by: gix-bot <gix-bot@users.noreply.github.com>
Co-authored-by: David de Kloet <122978264+dskloetd@users.noreply.github.com>
Co-authored-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: Max Strasinsky <98811342+mstrasinskis@users.noreply.github.com>