Stop using afterEach for cleanup (#5672)
# Motivation

Cleanup should be done in `beforeEach`, not in `afterEach`.
`afterEach` should only be used to verify expectations such as that
certain things did not happen during each test.

Remember: `beforeEach` is used for both cleanup and setup. Cleanup
should come before setup and while setup can be done in the `beforeEach`
of a `describe block that needs that setup, cleanup should always happen
in the top-level `beforeEach`.

# Changes

1. Just remove `afterEach` if the same cleanup already happens in
`beforeEach`.
2. Move contents of `afterEach` to `beforeEach`.
3. Change `afterEach` to `beforeEach` (if there isn't also a
`beforeEach`) and move it to the top level `describe` if it isn't
already.
4. In some cases, remove or merge `beforeAll`.
5. Remove `vi.clearAllMocks()` if there is also `vi.resetAllMocks()` as
the latter is strictly stronger.
6. Remove `someMock.mockClear()` if the mock is anyway newly created in
`beforeEach`.
7. Replace `vi.mock()` at the top of the file with `vi.spyOn()` in
`beforeEach` to avoid the 1-time mock being reset by
`vi.resetAllMocks()` in `beforeEach`.

# Tests

Changed tests still pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary