Add scripts/propose-install-canister (#5674)
# Motivation

When testing, it can be useful to replace one of the NNS canisters with
a custom built version.
This can be tricky because for some of the canisters we are not the
controller so they need to be installed via proposal, which requires a
lot of arguments to be constructed for a shell command.
It would be convenient to have a script that does this for you.

# Changes

1. Extract `scripts/get-neurons` from `scripts/vote` to get neurons of
an identity.
2. Add `scripts/propose-install-canister` which constructs and calls the
`ic-admin` command to propose installing a canister.

# Tests

Tested manually with
```
# Propose installing nns-dapp
scripts/propose-install-canister --canister nns-dapp --wasm nns-dapp.wasm.gz --arg-did-file nns-dapp-arg-local.did 

# Propose installing nns-governance
scripts/propose-install-canister --canister nns-governance --wasm governance-canister-test.wasm.gz 

# Propose installing nns-governance on DevEnv
scripts/propose-install-canister --canister nns-governance --wasm governance-canister-test.wasm.gz --identity devenv-snsdemo8 --network devenv_dskloet
```

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary